use hooks to separate-UI-and-logic
