--- app/components/editor/syntax.tsx
@@ -1,41 +1,0 @@
-// Copyright 2024 The casbin Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import React from 'react';
-import { Config } from 'casbin';
-
-interface SyntaxProps {
-  model: string;
-  onResponse: (com: JSX.Element) => void;
-}
-
-const Syntax = (props: SyntaxProps) => {
-  return (
-    <button
-      style={{ marginRight: 8 }}
-      onClick={() => {
-        try {
-          Config.newConfigFromText(props.model);
-          props.onResponse(<div>Passed</div>);
-        } catch (e) {
-          props.onResponse(<div>{(e as any).message}</div>);
-        }
-      }}
-    >
-      SYNTAX VALIDATE
-    </button>
-  );
-};
-
-export default Syntax;
